Added sigscan for other version of the game
The current ASL works for the 2 main versions of the game just fine, but if other version ever appears, this ASL version will work (in theory) and will be merged to main.

Did this as an experiment, and it worked, but I don't want to go out of my way to test it with other computers.